Various Fixes

+++ Implementation/data_structures/art/art.h
@@ -25,12 +25,12 @@ namespace art
         std::vector<uint32_t> FindRange(uint32_t from, uint32_t to) const;
 
         /**
-         * Prints the Tree in preorder.
+         * Prints the Tree in pre-order.
          */
         void PrintTree() const;
 
     private:
-        void ExpandLazyExpansion(uint32_t value1, uint32_t value2, uint8_t depth, Node* node);
+        void ExpandLazyExpansion(uint32_t value1, uint32_t value2, int depth, Node* node);
 
     private:
         Node* root_;
